+++ Source/Engine/Renderer/Renderer.h
@@ -18,7 +18,7 @@ namespace kiko
 		void BeginFrame();
 		void EndFrame();
 
-		void SetColor(int r, int g, int b, int a);
+		void SetColor(uint8_t r, uint8_t g, uint8_t b, uint8_t a);
 
 		void DrawLine(int x1, int y1, int x2, int y2);
 		void DrawLine(float x1, float y1, float x2, float y2);
